# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MiniAppUser {
  telegramId: number;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  username: string | null;
}

export interface MiniAppGameweek {
  id: number;
  /** @nullable */
  deadlineTime: string | null;
  locked: boolean;
  status: string;
}

export type MiniAppPlayerPosition = typeof MiniAppPlayerPosition[keyof typeof MiniAppPlayerPosition];


export const MiniAppPlayerPosition = {
  goalkeeper: 'goalkeeper',
  defender: 'defender',
  midfielder: 'midfielder',
  forward: 'forward',
} as const;

export interface MiniAppPlayer {
  id: number;
  name: string;
  club: string;
  clubId: number;
  position: MiniAppPlayerPosition;
  price: number;
  /** @nullable */
  status: string | null;
  /** @nullable */
  chanceOfPlayingThisRound: number | null;
  totalPoints: number;
  minutes: number;
}

export interface MiniAppTeam {
  selectedPlayerIds: number[];
  startingPlayerIds: number[];
  benchPlayerIds: number[];
  /** @nullable */
  captainPlayerId: number | null;
  /** @nullable */
  viceCaptainPlayerId: number | null;
  points: number;
  pointsSource: string;
  scoredPlayerIds: number[];
  /** @nullable */
  lastPointsUpdatedAt: string | null;
  submissionStatus: string;
  registered: boolean;
}

export interface MiniAppLeaderboardEntry {
  rank: number;
  displayName: string;
  points: number;
  isCurrentUser: boolean;
  /** @nullable */
  lastUpdatedAt: string | null;
}

export type MiniAppSignalItemKind = typeof MiniAppSignalItemKind[keyof typeof MiniAppSignalItemKind];


export const MiniAppSignalItemKind = {
  value: 'value',
  form: 'form',
  differential: 'differential',
  captain: 'captain',
  injury: 'injury',
  price_rise: 'price_rise',
  price_drop: 'price_drop',
} as const;

export type MiniAppSignalItemPosition = typeof MiniAppSignalItemPosition[keyof typeof MiniAppSignalItemPosition];


export const MiniAppSignalItemPosition = {
  goalkeeper: 'goalkeeper',
  defender: 'defender',
  midfielder: 'midfielder',
  forward: 'forward',
} as const;

export interface MiniAppSignalItem {
  kind: MiniAppSignalItemKind;
  playerId: number;
  playerName: string;
  position: MiniAppSignalItemPosition;
  club: string;
  price: number;
  title: string;
  detail: string;
}

export interface MiniAppSignal {
  available: boolean;
  message: string;
  disclaimer: string;
  signals: MiniAppSignalItem[];
}

export interface MiniAppBootstrap {
  user: MiniAppUser;
  gameweek: MiniAppGameweek;
  players: MiniAppPlayer[];
  team: MiniAppTeam;
  leaderboard: MiniAppLeaderboardEntry[];
  signal: MiniAppSignal;
}

export interface MiniAppTeamInput {
  /**
     * @minItems 15
     * @maxItems 15
     */
  selectedPlayerIds: number[];
  /**
     * @minItems 11
     * @maxItems 11
     */
  startingPlayerIds: number[];
  captainPlayerId: number;
  viceCaptainPlayerId: number;
}