# lib/api-zod/src/generated/types/miniAppBootstrap.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MiniAppGameweek } from './miniAppGameweek';
import type { MiniAppLeaderboardEntry } from './miniAppLeaderboardEntry';
import type { MiniAppPlayer } from './miniAppPlayer';
import type { MiniAppSignal } from './miniAppSignal';
import type { MiniAppTeam } from './miniAppTeam';
import type { MiniAppUser } from './miniAppUser';

export interface MiniAppBootstrap {
  user: MiniAppUser;
  gameweek: MiniAppGameweek;
  players: MiniAppPlayer[];
  team: MiniAppTeam;
  leaderboard: MiniAppLeaderboardEntry[];
  signal: MiniAppSignal;
}